+++ tests/organ_segmentation_test.py
@@ -184,30 +184,6 @@ and background")
         params = {}
         self.synthetic_liver_template(params)
 
-    def synthetic_liver(self):
-        """
-        Create synthetic data. There is some liver and porta -like object.
-        """
-        # data
-        slab = {'none': 0, 'liver': 1, 'porta': 2}
-        voxelsize_mm = np.array([1.0, 1.0, 1.2])
-
-        segm = np.zeros([80, 256, 250], dtype=np.int16)
-
-        # liver
-        segm[30:60, 70:180, 40:190] = slab['liver']
-        # porta
-        segm[40:45, 120:130, 70:190] = slab['porta']
-        segm[40:45, 80:130, 100:110] = slab['porta']
-        segm[40:44, 120:170, 130:135] = slab['porta']
-
-        data3d = np.zeros(segm.shape)
-        data3d[segm == slab['liver']] = 146
-        data3d[segm == slab['porta']] = 206
-        noise = (np.random.normal(0, 10, segm.shape))  # .astype(np.int16)
-        data3d = (data3d + noise).astype(np.int16)
-        return data3d, segm, voxelsize_mm, slab
-
     def synthetic_liver_template(self, params):
         """
         Function uses organ_segmentation  for synthetic box object
@@ -216,12 +192,12 @@ and background")
         # dcmdir = os.path.join(path_to_script,'./../sample_data/matlab/examples/sample_data/DICOM/digest_article/') # noqa
 # data
 
-        data3d, segm, voxelsize_mm, slab = self.synthetic_liver()
+        data3d, segm, voxelsize_mm, slab, seeds_liver, seeds_porta = io3d.datasets.generate_synthetic_liver()
 
 # seeds
-        seeds = np.zeros(data3d.shape, np.int8)
-        seeds[40:55, 90:120, 70:110] = 1
-        seeds[30:45, 190:200, 40:90] = 2
+#         seeds = np.zeros(data3d.shape, np.int8)
+#         seeds[40:55, 90:120, 70:110] = 1
+#         seeds[30:45, 190:200, 40:90] = 2
 # [mm]  10 x 10 x 10        # voxelsize_mm = [1, 4, 3]
         metadata = {'voxelsize_mm': voxelsize_mm}
 
@@ -229,7 +205,7 @@ and background")
             None,
             data3d=data3d,
             metadata=metadata,
-            seeds=seeds,
+            seeds=seeds_liver,
             working_voxelsize_mm=5,
             manualroi=False,
             autocrop=False,
@@ -239,8 +215,19 @@ and background")
 
         oseg.ninteractivity()
 
-        volume = oseg.get_segmented_volume_size_mm3()
+        volume = oseg.get_segmented_volume_size_mm3("liver")
+        self.assertGreater(volume, 570000, "Liver volume should have proper size")
+        self.assertLess(volume, 640000, "Liver volume should have proper size")
+
         oseg.portalVeinSegmentation(interactivity=False, threshold=180)
+
+        volume_porta = oseg.get_segmented_volume_size_mm3("porta")
+        self.assertGreater(volume_porta, 9000, "Liver volume should have proper size")
+        self.assertLess(volume_porta, 11000, "Liver volume should have proper size")
+        import sed3
+        ed = sed3.sed3(data3d,
+                       contour=(oseg.segmentation))
+        ed.show()
         oseg.saveVesselTree('porta')
 
         # print '> 0 '
@@ -253,16 +240,10 @@ and background")
         # print oseg.voxelsize_mm
 
         # import pdb; pdb.set_trace()
-        # import sed3
-        # ed = sed3.sed3(data3d, seeds=seeds,
-        #                contour=(oseg.segmentation))
-        # ed.show()
         # import ipdb; ipdb.set_trace() #  noqa BREAKPOINT
 
         # mel by to být litr. tedy milion mm3
         # je to zvlastni. pro nekter verze knihoven je to 630, pro jine 580
-        self.assertGreater(volume, 570000)
-        self.assertLess(volume, 640000)
 
     def test_roi(self):
         """